+++ packages/expo-modules-core/ios/Swift/Arguments/Convertibles.swift
@@ -9,6 +9,20 @@ import CoreGraphics
 /// As an example, when the `CGPoint` type is used as an argument type, its instance can be
 /// created from an array of two doubles or an object with `x` and `y` fields.
 
+// MARK: - Foundation
+
+extension URL: ConvertibleArgument {
+  public static func convert(from value: Any?) throws -> Self {
+    if let uri = value as? String, let url = URL(string: uri) {
+      // `URL(string:)` is an optional init but it doesn't imply it's a valid URL,
+      // so here we don't check for the correctness of the URL.
+      // If it has no scheme, we assume it was the file path.
+      return url.scheme != nil ? url : URL(fileURLWithPath: uri)
+    }
+    throw Conversions.ConvertingError<URL>(value: value)
+  }
+}
+
 // MARK: - UIKit
 
 extension UIColor: ConvertibleArgument {